+++ train_notify/core/__init__.py
@@ -1,1 +1,0 @@
-"""Core train_notify modules."""