+++ trainpulse/__init__.py
@@ -1,4 +1,4 @@
-"""Train Notify package."""
+"""TrainPulse package."""
 
 __all__ = ["__version__"]
 __version__ = "0.1.0"